+++ doc/CMakeLists.txt
@@ -40,15 +40,6 @@ project(documentation_sphinx)
 # message(STATUS "========Python executable: ${PYTHON_EXECUTABLE}")
 # message(STATUS "========PYTHON_VENV_PATH : ${PYTHON_VENV_PATH}")
 
-
-
-
-
-
-
-
-
-
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}")
 find_package(Sphinx REQUIRED)
 find_package(Doxygen REQUIRED)
@@ -74,7 +65,7 @@ add_custom_target(documentation_copy_sample_files ALL
   COMMAND cmake -E copy_directory ${ECAL_PROJECT_ROOT}/samples                                              ${DOC_SOURCE_DIRECTORY}/source_code_samples
   # blob samples
   COMMAND cmake -E copy_directory ${ECAL_PROJECT_ROOT}/ecal/samples/cpp/pubsub/binary                       ${DOC_SOURCE_DIRECTORY}/source_code_samples/cpp/binary/blob
-  # COMMAND cmake -E copy_directory ${ECAL_PROJECT_ROOT}/lang/c/samples/pubsub/blob                         ${DOC_SOURCE_DIRECTORY}/source_code_samples/c/binary/blob
+  COMMAND cmake -E copy_directory ${ECAL_PROJECT_ROOT}/lang/c/samples/pubsub/binary                         ${DOC_SOURCE_DIRECTORY}/source_code_samples/c/binary/blob
   COMMAND cmake -E copy_directory ${ECAL_PROJECT_ROOT}/lang/csharp/Eclipse.eCAL.Core.Samples/pubsub         ${DOC_SOURCE_DIRECTORY}/source_code_samples/csharp/binary/blob
   COMMAND cmake -E copy_directory ${ECAL_PROJECT_ROOT}/lang/python/samples/core/pubsub                      ${DOC_SOURCE_DIRECTORY}/source_code_samples/python/binary/blob
   # string samples
@@ -200,7 +191,3 @@ install(DIRECTORY ${SPHINX_HTML_DIR}/
   DESTINATION ${eCAL_install_doc_dir}
   COMPONENT doc 
 )
-
-if (ECAL_BUILD_DOCS_SAMPLES)
-  add_subdirectory(samples)
-endif()
