--- doc/samples/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_subdirectory(cpp/string/hello_world_snd)
-add_subdirectory(cpp/string/hello_world_rec)
-add_subdirectory(cpp/protobuf/hello_world_snd)
-add_subdirectory(cpp/protobuf/hello_world_rec)
-add_subdirectory(cpp/configuration/hello_config)
-add_subdirectory(cpp/configuration/publisher_config)
