--- doc/samples/cpp/configuration/hello_config/CMakeLists.txt
@@ -1,19 +1,0 @@
-cmake_minimum_required(VERSION 3.16)
-
-project(hello_config)
-
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-find_package(eCAL REQUIRED)
-
-set(source_files
-  main.cpp
-)
-
-add_executable(${PROJECT_NAME} ${source_files})
-
-target_link_libraries(${PROJECT_NAME}
-  PRIVATE
-  eCAL::string_core
-)
