--- doc/samples/cpp/protobuf/hello_world_snd/CMakeLists.txt
@@ -1,26 +1,0 @@
-cmake_minimum_required(VERSION 3.16)
-
-project(protobuf_snd)
-
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-find_package(eCAL REQUIRED)
-find_package(Protobuf REQUIRED)
-
-set(source_files
-  main.cpp
-)
-
-set(protobuf_files
-    ${CMAKE_CURRENT_SOURCE_DIR}/../proto_messages/hello_world.proto
-)
-
-add_executable(${PROJECT_NAME} ${source_files})
-
-PROTOBUF_TARGET_CPP(${PROJECT_NAME} ${CMAKE_CURRENT_SOURCE_DIR}/../proto_messages/ ${protobuf_files})
-
-target_link_libraries(${PROJECT_NAME}
-  PRIVATE
-  eCAL::protobuf_core
-)
