--- doc/samples/cpp/protobuf/proto_messages/hello_world.proto
@@ -1,10 +1,0 @@
-syntax = "proto3";
-
-package proto_messages;
-
-message HelloWorld
-{
-  string name      = 1;
-  uint32 id        = 2;
-  string msg       = 3;
-}
