--- doc/samples/cpp/string/hello_world_rec/main.cpp
@@ -1,32 +1,0 @@
-#include <ecal/ecal.h>
-#include <ecal/msg/string/subscriber.h>
-
-#include <iostream>
-#include <thread>
-
-// Callback for receiving messages
-void HelloWorldCallback(const std::string& message)
-{
-  std::cout << "Received Message: " << message << std::endl;
-}
-
-int main(int argc, char** argv)
-{
-  // Initialize eCAL
-  eCAL::Initialize("Hello World Subscriber");
-
-  // Create a subscriber that listenes on the "hello_world_topic"
-  eCAL::string::CSubscriber subscriber("hello_world_topic");
-
-  // Set the Callback
-  subscriber.SetReceiveCallback(std::bind(&HelloWorldCallback, std::placeholders::_2));
-
-  // Just don't exit
-  while (eCAL::Ok())
-  {
-    std::this_thread::sleep_for(std::chrono::milliseconds(500));
-  }
-
-  // finalize eCAL API
-  eCAL::Finalize();
-}
