+++ lang/csharp/CMakeLists.txt
@@ -33,6 +33,8 @@ set(ECAL_CSHARP_PROTOBUF_VERSION "${Protobuf_VERSION}")
 endif ()
 endmacro()
 
+set(CMAKE_DOTNET_TARGET_FRAMEWORK_VERSION  "v4.7.1")
+
 option(ECAL_CSHARP_BUILD_SAMPLES "BUILD eCAL C# samples" ON)
 option(ECAL_CSHARP_BUILD_TESTS   "BUILD eCAL C# unittests" ON)
