+++ lang/csharp/Eclipse.eCAL.Protobuf.Datatypes/CMakeLists.txt
@@ -42,7 +42,6 @@ PROTOBUF_TARGET_CSHARP(${PROJECT_NAME} ${PROTOBUF_ORIGIN_DIR}
 
 set_target_properties(${PROJECT_NAME} PROPERTIES
     VS_GLOBAL_ROOTNAMESPACE ${PROJECT_NAME}
-    VS_DOTNET_TARGET_FRAMEWORK_VERSION "v4.7.1"
     VS_DOTNET_REFERENCES "System"
 )
