+++ lang/csharp/Eclipse.eCAL.Protobuf.Samples/pubsub/person_send_csharp.cs
@@ -18,38 +18,37 @@
 */
 
 /**
- * @file hello_send.cs
+ * @file person_snd_csharp.cs
  *
- * @brief A minimal example of using the eCAL API to send string messages.
+ * @brief A minimal example of using the eCAL API to send protobuf messages.
  *
  * This example demonstrates how to initialize the eCAL API, print version information,
- * create a string publisher for the topic "hello", construct and send messages, and keep
- * the application running until eCAL is terminated. It serves as a basic reference for
- * implementing a string publisher in C#.
+ * create a protobuf publisher for the topic "person" (using "Person" as the data type), construct
+ * and send messages, and keep the application running until eCAL is terminated. It serves as
+ * a basic reference for implementing a protobuf publisher in C#.
  */
 
 using System;
-using System.Threading;
 // Include the eCAL API namespace
-using Continental.eCAL.Core;
+using Eclipse.eCAL.Core;
 
-public class HelloSend
+public class PersonSend
 {
-  public static void Main()
+  static void Main()
   {
-    Console.WriteLine("------------------------");
-    Console.WriteLine(" C#: HELLO WORLD SENDER");
-    Console.WriteLine("------------------------");
+    Console.WriteLine("-------------------");
+    Console.WriteLine(" C#: PERSON SENDER");
+    Console.WriteLine("-------------------");
 
     /*
       Initialize eCAL. You always have to initialize eCAL before using its API.
-      The name of our eCAL Process will be "hello_send_csharp". 
+      The name of our eCAL Process will be "person send c#". 
       This name will be visible in the eCAL Monitor, once the process is running.
     */
-    Core.Initialize("hello_send_csharp");
+    Core.Initialize("person send c#");
 
     /*
-      Print eCAL version information.
+      Print version info.
     */
     Console.WriteLine(String.Format("eCAL {0} ({1})\n", Core.GetVersion(), Core.GetDate()));
 
@@ -61,32 +60,47 @@ public class HelloSend
     // This function is not wrapped yet.
 
     /*
-      Now we create a new publisher that will publish the topic "hello".      
+      Now we create a new publisher that will publish the topic "person".
+      The data type is "Pb.People.Person", generated from the protobuf definition.    
     */
-    StringPublisher publisher = new StringPublisher("hello");
+    var publisher = new ProtobufPublisher<Pb.People.Person>("person");
 
+    /*
+      Construct a message. The message is a protobuf struct that will be sent to the subscribers.
+    */
+    var person = new Pb.People.Person
+    {
+      Id = 0,
+      Email = "max@online.de",
+      Name = "Max",
+      Stype = Pb.People.Person.Types.SType.Female,
+      Dog = new Pb.Animal.Dog { Name = "Brandy" },
+      House = new Pb.Environment.House { Rooms = 4 }
+    };
+
+    int loop_count = 0;
     /*
       Creating an infinite publish-loop.
       eCAL Supports a stop signal; when an eCAL Process is stopped, eCAL_Ok() will return false.
     */
-    int loop_count = 0;
     while (Core.Ok())
     {
       /*
-        Construct a message. The message is a string that will be sent to the subscribers.
+        Change in each loop the content of the message to see a difference per message.
       */
-      string message = String.Format("HELLO WORLD FROM C# {0,6}", ++loop_count);
+      person.Id = loop_count;
+      loop_count++;
 
       /*
-        Send the message. The message is sent to all subscribers that are currently connected to the topic "hello".
+        Send the message. The message is sent to all subscribers that are currently connected to the topic "person".
       */
-      publisher.Send(message);
-      Console.WriteLine(String.Format("Sent: {0}", message));
+      publisher.Send(person);
+      Console.WriteLine(String.Format("Sending: {0}", person.ToString()));
 
       /*
         Sleep for 500ms to send in a frequency of 2 hz.
       */
-      Thread.Sleep(500);
+      System.Threading.Thread.Sleep(500);
     }
 
     /*
