+++ lang/csharp/Eclipse.eCAL.Protobuf/CMakeLists.txt
@@ -1,6 +1,6 @@
 # ========================= eCAL LICENSE =================================
 #
-# Copyright (C) 2016 - 2019 Continental Corporation
+# Copyright (C) 2016 - 2025 Continental Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -32,7 +32,6 @@ target_link_libraries(${PROJECT_NAME} PUBLIC Eclipse.eCAL.Core)
 
 set_target_properties(${PROJECT_NAME} PROPERTIES
   VS_GLOBAL_ROOTNAMESPACE ${PROJECT_NAME}
-  VS_DOTNET_TARGET_FRAMEWORK_VERSION "v4.7.1"
 )
 
 if (ECAL_CSHARP_PROTOBUF_VERSION VERSION_GREATER "3.16.0")
