+++ lang/python/samples/core/pubsub/blob_receive.py
@@ -35,9 +35,9 @@ def main():
   print("-----------------------")
   
   # Initialize eCAL. You always have to initialize eCAL before using it.
-  # The name of our eCAL Process will be "blob_receive_python".
+  # The name of our eCAL Process will be "blob receive python".
   # This name will be visible in the eCAL Monitor, once the process is running.
-  ecal_core.initialize("blob_receive_python")
+  ecal_core.initialize("blob receive python")
   
   # Print used eCAL version and date
   print("eCAL {} ({})\n".format(ecal_core.getversion(), ecal_core.getdate()))
