+++ lang/python/samples/core/pubsub/hello_send.py
@@ -29,9 +29,9 @@ def main():
   print("-------------------------------")
   
   # Initialize eCAL. You always have to initialize eCAL before using it.
-  # The name of our eCAL Process will be "hello_send_python".
+  # The name of our eCAL Process will be "hello send python".
   # This name will be visible in the eCAL Monitor, once the process is running.
-  ecal_core.initialize("hello_send_python")
+  ecal_core.initialize("hello send python")
 
   # Print used eCAL version and date
   print("eCAL {} ({})\n".format(ecal_core.getversion(), ecal_core.getdate()))
