+++ lang/python/samples/core/pubsub/person_send.py
@@ -35,9 +35,9 @@ def main():
   print("-----------------------")
   
   # Initialize eCAL. You always have to initialize eCAL before using it.
-  # The name of our eCAL Process will be "person_send_python".
+  # The name of our eCAL Process will be "person send python".
   # This name will be visible in the eCAL Monitor, once the process is running.
-  ecal_core.initialize("person_send_python")
+  ecal_core.initialize("person send python")
   
   # Print used eCAL version and date
   print("eCAL {} ({})\n".format(ecal_core.getversion(), ecal_core.getdate()))
